Fix docstrings to match effective Go convention
Further to pull request in #75.

Turned out to be a surprisingly large amount of low value work :(. It's done now I suppose.

This was done on a best effort basis. I've reviewed the patch and don't think there are any non-comment/whitespace changes present. I have verified this by grepping for `^\+[^/]` which looks for diff lines not starting with /.

There was one minor thing where I reordered a small function to make it easier to edit the comments in bulk.